requirements.txt
doco==0.0.3
Django==1.11.5
django-bootstrap-form==3.3
django-debug-toolbar==1.8
django-toolbelt==0.0.1
entrypoints==0.2.2
findspark==1.1.0
future==0.16.0
ggplot==0.11.5
gunicorn==19.7.1
holoviews==1.8.4
html5lib==0.999999999
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipykernel==4.6.1
ipython==5.3.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
lancet-ioam==0.9.0
line-bot-sdk==1.5.0
lxml==4.0.0
MarkupSafe==1.0
matplotlib==2.0.0
mistune==0.7.4
nbconvert==5.1.1
nbformat==4.3.0
numpy==1.13.3
oauthlib==2.0.4
olefile==0.44
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
param==1.5.1
plotly==2.1.0
prompt-toolkit==1.0.14
pya3rt==1.1
python-dateutil==2.6.0
qtconsole==4.3.0
queuelib==1.4.2
requests==2.18.4
requests-oauthlib==0.8.0
retrying==1.3.3
scikit-learn==0.19.0
scipy==0.19.1
scrapinghub==2.0.1
Scrapy==1.4.0
seaborn==0.7.1
service-identity==17.0.0
shub==2.7.0
simplegeneric==0.8.1
simplejson==3.11.1
six==1.11.0
smart-open==1.5.3
sqlparse==0.2.4
static3==0.7.0
statsmodels==0.8.0
Steganocrypto==0.0.3
stevedore==1.27.0
terminado==0.6
testpath==0.3
tflearn==0.3.2
tornado==4.4.3
tqdm==4.19.4
traitlets==4.3.2
urllib3==1.22
w3lib==1.18.0